move node test utils to e2e package
