+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/tests/electron-code-aad.spec.ts
@@ -8,23 +8,21 @@ import {
     Browser,
 } from "@playwright/test";
 
-import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
+import { NodeCacheTestUtils } from "e2e-test-utils/src/NodeCacheTestUtils";
 import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
 import { AppTypes, AzureEnvironments } from "e2e-test-utils/src/Constants";
 import { LabClient } from "e2e-test-utils/src/LabClient";
-import { setupCredentials } from "e2e-test-utils/src/TestUtils";
+import { 
+    setupCredentials, 
+    SCREENSHOT_BASE_FOLDER_NAME, 
+    validateCacheLocation 
+} from "e2e-test-utils/src/TestUtils";
 
 import {
     Screenshot,
     enterCredentials,
     retrieveAuthCodeUrlFromBrowserContext,
 } from "e2e-test-utils/src/ElectronPlaywrightTestUtils";
-
-import {
-    SCREENSHOT_BASE_FOLDER_NAME,
-    validateCacheLocation,
-} from "../../testUtils";
-
 import * as path from "path";
 
 let electronApp: ElectronApplication;
