+++ samples/msal-node-samples/auth-code/test/auth-code-adfs.spec.ts
@@ -1,15 +1,18 @@
 import * as puppeteer from "puppeteer";
-import {Screenshot, createFolder, setupCredentials, RETRY_TIMES} from "e2e-test-utils/src/TestUtils";
-import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
-import { LabClient } from "e2e-test-utils/src/LabClient";
-import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
-import { AppTypes, AzureEnvironments, FederationProviders, UserTypes } from "e2e-test-utils/src/Constants";
 import {
+    Screenshot, 
+    createFolder, 
+    setupCredentials, 
+    RETRY_TIMES, 
     enterCredentialsADFS,
     enterCredentialsADFSWithConsent,
     SCREENSHOT_BASE_FOLDER_NAME,
-    SAMPLE_HOME_URL,
- } from "../../testUtils";
+    SAMPLE_HOME_URL
+} from "e2e-test-utils/src/TestUtils";
+import { NodeCacheTestUtils } from "e2e-test-utils/src/NodeCacheTestUtils";
+import { LabClient } from "e2e-test-utils/src/LabClient";
+import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
+import { AppTypes, AzureEnvironments, FederationProviders, UserTypes } from "e2e-test-utils/src/Constants";
 import { PublicClientApplication } from "@azure/msal-node";
 
 const TEST_CACHE_LOCATION = `${__dirname}/data/adfs.cache.json`;
