+++ samples/msal-node-samples/device-code/test/device-code-aad-agc.spec.ts
@@ -4,15 +4,17 @@
  */
 
 import * as puppeteer from "puppeteer";
-import {Screenshot, createFolder, RETRY_TIMES} from "e2e-test-utils/src/TestUtils";
-import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
 import {
+    Screenshot, 
+    createFolder, 
+    RETRY_TIMES, 
     approveRemoteConnect,
     enterCredentials,
     enterDeviceCode,
     SCREENSHOT_BASE_FOLDER_NAME,
     validateCacheLocation
- } from "../../testUtils";
+} from "e2e-test-utils/src/TestUtils";
+import { NodeCacheTestUtils } from "e2e-test-utils/src/NodeCacheTestUtils";
 import { Configuration, PublicClientApplication } from "@azure/msal-node";
 import { getKeyVaultSecretClient, getCredentials } from "e2e-test-utils/src/KeyVaultUtils";
