+++ samples/msal-node-samples/device-code/test/device-code-aad.spec.ts
@@ -4,19 +4,20 @@
  */
 
 import * as puppeteer from "puppeteer";
-import {Screenshot, createFolder, setupCredentials, RETRY_TIMES} from "e2e-test-utils/src/TestUtils";
-import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
-import { LabClient } from "e2e-test-utils/src/LabClient";
-import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
-import { AppTypes, AzureEnvironments } from "e2e-test-utils/src/Constants";
 import {
+    Screenshot, 
+    createFolder, 
+    setupCredentials, 
+    RETRY_TIMES, 
     approveRemoteConnect,
     enterCredentials,
     enterDeviceCode,
     SCREENSHOT_BASE_FOLDER_NAME,
-    validateCacheLocation
- } from "../../testUtils";
-
+    validateCacheLocation} from "e2e-test-utils/src/TestUtils";
+import { NodeCacheTestUtils } from "e2e-test-utils/src/NodeCacheTestUtils";
+import { LabClient } from "e2e-test-utils/src/LabClient";
+import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
+import { AppTypes, AzureEnvironments } from "e2e-test-utils/src/Constants";
 import { Configuration, PublicClientApplication } from "@azure/msal-node";
 
 // Set test cache name/location
