+++ samples/msal-node-samples/silent-flow/test/silent-flow-b2c-local.spec.ts
@@ -10,20 +10,17 @@ import {
     setupCredentials,
     b2cLocalAccountEnterCredentials,
     ONE_SECOND_IN_MS,
-    RETRY_TIMES
-} from "e2e-test-utils/src/TestUtils";
-import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
-import { LabClient } from "e2e-test-utils/src/LabClient";
-import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
-import { B2cProviders, UserTypes } from "e2e-test-utils/src/Constants";
-import {
+    RETRY_TIMES,
     clickSignIn,
     SCREENSHOT_BASE_FOLDER_NAME,
     SAMPLE_HOME_URL,
     SUCCESSFUL_GET_ALL_ACCOUNTS_ID,
     validateCacheLocation
-} from "../../testUtils";
-
+} from "e2e-test-utils/src/TestUtils";
+import { NodeCacheTestUtils } from "e2e-test-utils/src/NodeCacheTestUtils";
+import { LabClient } from "e2e-test-utils/src/LabClient";
+import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
+import { B2cProviders, UserTypes } from "e2e-test-utils/src/Constants";
 import { PublicClientApplication, TokenCache } from "@azure/msal-node";
 
 // Set test cache name/location
